+++ apps/tuk-web/src/app/invite/meet/[meetId]/page.tsx
@@ -1,5 +1,5 @@
-import InviteProposalContent from '@/app/invite/meet/[meetId]/src/components/InviteProposalContent';
+import InviteProposal from '@/app/invite/meet/[meetId]/src/components/InviteProposal';
 
 export default function InviteProposalPage() {
-  return <InviteProposalContent />;
+  return <InviteProposal />;
 }
